Add AR solid cards recognition project
