+++ ARSolidCardsRecognition/ARSolidCardsRecognition/Assets.xcassets/AR Resources.arresourcegroup/SOLID-dependency-injection.arreferenceimage/Contents.json
@@ -1,0 +1,16 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "Dependency-Inversion-Principle_3502x2385.png"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "properties" : {
+    "width" : 15,
+    "unit" : "centimeters"
+  }
+}
