+++ ARSolidCardsRecognition/ARSolidCardsRecognition/Assets.xcassets/AR Resources.arresourcegroup/SOLID-open-close.arreferenceimage/Contents.json
@@ -1,0 +1,15 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "open_close_princile_3502x2385.png"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "properties" : {
+    "width" : 15
+  }
+}
