+++ ARSolidCardsRecognition/ARSolidCardsRecognition/Assets.xcassets/AR Resources.arresourcegroup/SOLID-segregation.arreferenceimage/Contents.json
@@ -1,0 +1,16 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "Interface-Segregation-Principle_3502x2385.png"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "properties" : {
+    "width" : 15,
+    "unit" : "centimeters"
+  }
+}
