+++ ARSolidCardsRecognition/ARSolidCardsRecognition/Assets.xcassets/AR Resources.arresourcegroup/SOLID-substitution-principle.arreferenceimage/Contents.json
@@ -1,0 +1,15 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "Liskov's-substitution-principle_3502x2385.png"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "properties" : {
+    "width" : 15
+  }
+}
