+++ ARSolidCardsRecognition/ARSolidCardsRecognition/Assets.xcassets/AR Resources.arresourcegroup/car.arreferenceimage/Contents.json
@@ -1,0 +1,15 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "IMG_5634.png"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "properties" : {
+    "width" : 1
+  }
+}
